Backend/legal-rag-backend/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
vertexai
google-cloud-documentai
python-dotenv
numpy
python-multipart
langchain


python-multipart

# CORS middleware
starlette
fastapi
pinecone
python-dotenv
google-cloud-documentai
vertexai
numpy
PyMuPDF
google-cloud-firestore
requests
langchain-text-splitters
weasyprint
jinja2
pytz
matplotlib
vertexai
google-cloud-aiplatform
python-multipart
google-adk
presidio-analyzer
presidio-anonymizer
spacy
pdfplumber
reportlab